[SPARK-43043][CORE] Improve the performance of MapOutputTracker.updateMapOutput
### What changes were proposed in this pull request?

The PR changes the implementation of MapOutputTracker.updateMapOutput() to search for the MapStatus under the help of a mapping from mapId to mapIndex, previously it was performing a linear search, which would become performance bottleneck if a large proportion of all blocks in the map are migrated.

### Why are the changes needed?

To avoid performance bottleneck when block decommission is enabled and a lot of blocks are migrated within a short time window.

### Does this PR introduce _any_ user-facing change?

No, it's pure performance improvement.

### How was this patch tested?

Manually test.

Closes #40690 from jiangxb1987/SPARK-43043.

Lead-authored-by: Xingbo Jiang <xingbo.jiang@databricks.com>
Co-authored-by: Jiang Xingbo <jiangxb1987@gmail.com>
Signed-off-by: Xingbo Jiang <xingbo.jiang@databricks.com>

(cherry picked from commit 66a2eb8f8957c22c69519b39be59beaaf931822b)