[SPARK-44718][SQL] Match ColumnVector memory-mode config default to OffHeapMemoryMode config value
### What changes were proposed in this pull request?

Set the column vector default memory mode to depend on the off-heap memory mode flag. This is to prevent a user from using Vectorized-Reader with an on-heap column-vector by default when the off-heap memory mode is enabled on the cluster.

### Why are the changes needed?
Avoid the unintentional usage of on-heap memory in vectorized-reader when off-heap memory mode is enabled by the user.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Manual & existing tests.

Closes #42394 from majdyz/offheap-colvec-mode-default-value.

Lead-authored-by: Zamil Majdy <zamil.majdy@databricks.com>
Co-authored-by: Zamil Majdy <zamil.majdy@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit afcccb42c96cc7785a85c9463523f34d7a900a1d)