[SPARK-48761][SQL] Introduce clusterBy DataFrameWriter API for Scala
Introduce a new `clusterBy` DataFrame API in Scala. This PR adds the API for both the DataFrameWriter V1 and V2, as well as Spark Connect.

Introduce more ways for users to interact with clustered tables.

Yes, it adds a new `clusterBy` DataFrame API in Scala to allow specifying the clustering columns when writing DataFrames.

New unit tests.

No

Closes #47301 from zedtang/clusterby-scala-api.

Authored-by: Jiaheng Tang <jiaheng.tang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>